Rename optimize-fs-ops module to optimize-mkdirs
