+++ src/util/optimize-mkdirs.ts
